注册和搜索好友

+++ interface/src/main/java/com/haley/water/dto/RankingListDto.java
@@ -31,6 +31,6 @@ public class RankingListDto {
     /**
      * 用水量
      */
-    private Integer waterConsumption;
+    private double waterConsumption;
 
 }
